requirements.txt
click==8.0.1
cycler==0.10.0
Flask==2.0.2
ftfy==6.0.3
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pyparsing==2.4.7
#pytesseract @ file:///home/jamie/pan-card-extract/pytesseract
python-dateutil==2.8.2
six==1.16.0
wcwidth==0.2.5
Werkzeug==2.0.2
